Correcao de alguns exercicios

+++ modulo2/ex10/sumIntValuesCarry.h
@@ -1,4 +1,4 @@
 #ifndef SUMINTVALUESCARRY_H
 #define SUMINTVALUESCARRY_H
-	short sumIntValuesCarry();
+	int sumIntValuesCarry();
 #endif
